Simple variational architecture that allows the decoder to sample from the posterior.
Briefly, the architecture is akin to conditional variational autoencoder (CVAE)-based transformer. There are two encoders: one for encoding the conditional features (actual Protoscribe embeddings) and one similar to the encoder in VAE (forming the latent space).

The latents produced by the encoder are blended with either the encoded
conditional features or with the embeddings of decoder input tokens.

PiperOrigin-RevId: 783885053